Change the way we install ruby
Prior to this commit, we installed ruby-install, then used ruby-install
to install ruby as root (but not as the system ruby), and installed a
few gems necessary for using homesick as root. Now, we install
ruby-install and chruby as root, but use ruby-install to install the
latest version of ruby as the USER dev.